+++ saito-analytics/src/analyse.rs
@@ -23,21 +23,8 @@ use saito_core::common::defs::{
     Currency, SaitoHash, SaitoPrivateKey, SaitoPublicKey, SaitoSignature, SaitoUTXOSetKey,
     Timestamp, UtxoSet, GENESIS_PERIOD, MAX_STAKER_RECURSION,
 };
-
-fn read_block(path: String) -> io::Result<Block> {
-    let bytes = match fs::read(&path) {
-        Ok(bytes) => bytes,
-        Err(e) => {
-            eprintln!("Failed to read file: {}", e);
-            return Err(e);
-        }
-    };
-
-    let deserialized_block = Block::deserialize_from_net(bytes)
-        .map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))?;
-
-    Ok(deserialized_block)
-}
+//mod sutils;
+use crate::sutils::get_blocks;
 
 fn analyse_block(block: Block) {
     //println!("deserialized_block {:?}" , deserialized_block);
@@ -50,43 +37,6 @@ fn analyse_block(block: Block) {
     // println!("{:?}" , tx);
     // println!("{:?}" , tx.timestamp);
     // println!("amount: {:?}" , tx.to[0].amount);
-}
-
-//read block directory as block vector
-fn get_blocks(directory_path: &str) -> io::Result<Vec<Block>> {
-    let mut blocks = Vec::new();
-
-    for entry in fs::read_dir(directory_path)? {
-        let entry = entry?;
-        let path = entry.path();
-
-        if path.is_file() {
-            match path.into_os_string().into_string() {
-                Ok(path_string) => {
-                    match read_block(path_string) {
-                        Ok(block) => blocks.push(block),
-                        Err(e) => {
-                            eprintln!("Failed to read block: {}", e);
-                            continue;
-                        }
-                    };
-                }
-                Err(_) => println!("Path contains non-unicode characters"),
-            }
-        }
-    }
-
-    blocks.sort_by(|a, b| {
-        if a.id < b.id {
-            Ordering::Less
-        } else if a.id > b.id {
-            Ordering::Greater
-        } else {
-            Ordering::Equal
-        }
-    });
-
-    Ok(blocks)
 }
 
 pub fn runAnalytics() {
@@ -105,7 +55,6 @@ pub fn runAnalytics() {
         Ok(blocks) => {
             println!("Got {} blocks", blocks.len());
             for block in blocks {
-                // Now you can analyse each block
                 analyse_block(block.clone());
             }
         }
